DELETEの例外ハンドリング実装

+++ src/main/java/com/example/usecase/EmployeeServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.usecase;
 
+import com.example.Exception.EmployeesNotFoundException;
 import com.example.domain.RequestEmployee;
 import com.example.infrastructure.EmployeesMapper;
 import lombok.AllArgsConstructor;
@@ -8,6 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.beans.Transient;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @AllArgsConstructor
@@ -35,6 +37,10 @@ public class EmployeeServiceImpl implements EmployeeService{
     @Override
     @Transactional
     public void delete(String id) {
+        int count = employeesMapper.delete(id);
+        if (count == 0) {
+            throw new EmployeesNotFoundException("specified employee [id = " + id + "] is not found.");
+        }
         employeesMapper.delete(id);
     }
 }
